backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
sentence-transformers==2.7.0
networkx==3.3
numpy==1.26.4
scipy==1.13.0
httpx==0.27.0
arxiv==2.1.0
google-generativeai==0.7.2
structlog==24.2.0
aiofiles==23.2.1
python-multipart==0.0.9
anthropic>=0.30.0
openai>=1.30.0
